feat: introduce vite plugin
