+++ packages/remix-routes/package.json
@@ -8,13 +8,15 @@
   },
   "files": [
     "lib",
+    "vite.js",
+    "vite.d.js",
     "!lib/__tests__"
   ],
   "scripts": {
     "watch": "tsc --watch",
     "prebuild": "rm -rf lib",
     "build": "tsc",
-    "prepare": "npm run build",
+    "prepack": "npm run build",
     "semantic-release": "semantic-release"
   },
   "keywords": [
@@ -25,7 +27,7 @@
   "author": "Wei Zhu",
   "license": "MIT",
   "devDependencies": {
-    "@remix-run/dev": "^1.16.1",
+    "@remix-run/dev": "^2.6.0",
     "@types/cli-table": "^0.3.0",
     "@types/ejs": "^3.1.2",
     "@types/jest": "^27.0.3",
@@ -33,7 +35,8 @@
     "@types/node": "^16.11.12",
     "@types/pluralize": "^0.0.29",
     "semantic-release": "^19.0.3",
-    "typescript": "^4.5.3"
+    "typescript": "^4.5.3",
+    "vite": "^3.1.0"
   },
   "dependencies": {
     "chokidar": "^3.5.2",
