+++ packages/remix-routes/vite.d.ts
@@ -1,0 +1,1 @@
+export declare function remixRoutes(): Vite.Plugin[];
