+++ packages/remix-routes/vite.js
@@ -1,0 +1,1 @@
+module.exports = require('./lib/vite');
